pyproject.toml
[project]
name = "chatterbox-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts>=0.1.3",
    "numpy>=2.4.2",
    "pkuseg==0.0.25",
    "setuptools==80.8.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
]
